move widgets to appropriate folder
