+++ lib/screens/memoryFeedScreen.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:mapbox_maps_example/point_annotations_example.dart';
-import 'BottomIconBar.dart';
-import 'Memory_item.dart';
-import 'Memory_header.dart';
-import 'Memory_content.dart';
-import 'Memory_reactions.dart';
-import 'SearchBarWidget.dart';
+import '../widgets/BottomIconBar.dart';
+import '../widgets/Memory_item.dart';
+import '../widgets/Memory_header.dart';
+import '../widgets/Memory_content.dart';
+import '../widgets/Memory_reactions.dart';
+import '../widgets/SearchBarWidget.dart';
 
 class MemoryFeedScreen extends StatefulWidget {
   const MemoryFeedScreen({Key? key}) : super(key: key);
