+++ lib/widgets/Memory_content.dart
