+++ lib/widgets/SearchBarWidget.dart
